Отзывы

+++ src/main/java/ru/yandex/practicum/filmorate/dal/mappers/ReviewRowMapper.java
@@ -13,8 +13,8 @@ public class ReviewRowMapper implements RowMapper<Review> {
                 .reviewId(rs.getLong("review_id"))
                 .content(rs.getString("content"))
                 .isPositive(rs.getBoolean("is_positive"))
-                .userId(rs.getInt("user_id"))
-                .filmId(rs.getInt("film_id"))
+                .userId(rs.getLong("user_id"))
+                .filmId(rs.getLong("film_id"))
                 .useful(rs.getInt("useful"))
                 .timestamp(rs.getTimestamp("created_at").getTime())
                 .build();
