+++ src/main/java/ru/yandex/practicum/filmorate/model/Review.java
@@ -9,8 +9,8 @@ public class Review {
     private Long reviewId;
     private String content;
     private Boolean isPositive;
-    private Integer userId;
-    private Integer filmId;
+    private Long userId;
+    private Long filmId;
     private Integer useful;
     private Long timestamp;
 }
